requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-groq
groq
python-multipart
aiofiles
langchain-core